Changes:
 - Performance issues were due to two reasons: long timeouts on libtpms
and low request timeout for cassandra in vvdproc. This causes many
insert statements to timeout because by using BufferedReader we are able
to read a lot more data and quickly, such that we basically end up
hammering Cassandra, which has a hard time keeping up. Increasing the
timeout solved this.
 - Further performance improvements from increasing the maximum
pipe-size to 32 mb and setting that to be the size of our vvdump pipe.
 - Some improvements due to using a small buffer size in BufferedReader,
which causes vvdproc to read from the pipe more often, ensuring that
there is space for the fuzzed processes to write to it.
 - Removed parallel fuzzing code; no need for this because libtpms is
running fast now.
 - Updated Cassandra configs to include password.
 - Outputting more progress statistics from vvdproc.